lcg_rand() is non-negative and int32

+++ src/lcg.h
@@ -2,5 +2,5 @@
 
 #include <cstdint>
 
-void lcg_srand(uint64_t seed);
+void lcg_srand(int seed);
 int lcg_rand();
